Implemented Contract pattern in fragments
Implemented the contract pattern in fragments so every event fired in then is handled by the parent activity. With this the clicks on the characters and houses in the RecyclerViews are handled by the activities instead of the adapters directly. Using this pattern will allow us to use a multi pane architecture in a better way.
GridView of house detail replaced by a RecyclerView with GridLayoutManager. Also develop and use a fragment in this activity.
Major refactor of code and updated UI tests